Cherry-pick performance improvements from AIRFLOW-2203
The first change avoids 4 calls to `configuration.getint()` (which reads
from the filesystem) each time we instantiate an Operator.

The second change defers cycle-detection until the DAG is evaluated
instead of eagerly checking for a cycle every time a task is added
during the setup phase.

Initial testing on TARS indicated a ~14% speedup when loading the full
DagBag:

- Before: 191 seconds
- After: 165 seconds

Commit 1:

[AIRFLOW-2203] Cherry-pick change to speed up Operator.__init__

Cherry-pick of this commit from upstream:
https://github.com/apache/airflow/commit/92357d53e60a6c88afe6792923ff68070ec81df3

Set default values of Resources loaded from the configuration to prevent
4x Config lookups for every task created.

Commit 2:

[AIRFLOW-2203] Defer cycle detection

Cherry-pick of this commit from upstream:
https://github.com/apache/airflow/commit/c3730650c852cd7a5e06a5933f5064bbb04e0e88
Moved from adding_task to when dag is being bagged.
This changes import dag runtime from polynomial to somewhat linear.