Always reload DAGs when running sync_to_db (#233)
If the DagBag is initialized with `store_serialized_dags = True` then it
skips evaluating the DAG definitions on disk. We always want to eval the
DAG definitions when we run `airflow sync_to_db` - otherwise this
command is effectively a no-op.

This change explicitly sets `store_serialized_dags = False` when
initializing the DagBag so that we'll reload DAG definitions before
serializing and storing them in the database.