Instrument duration of scheduler loop
According to [1], sensors with `mode=reschedule` break when the duration
of the scheduler loop >= the reschedule interval.

Let's add some instrumentation to identify the maximum time we spend in
the scheduling loop, as this number (plus some margin of safety)  will
be the lower bound on the reschedule interval until the race condition
is fixed.

[1] https://issues.apache.org/jira/browse/AIRFLOW-5071?focusedCommentId=17206032&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-17206032