[AIRFLOW-5071] Fix for race condition with mode=reschedule
This is broken upstream - see
https://issues.apache.org/jira/browse/AIRFLOW-5071

This fix feels like a hack, so I'm not sure that it makes a good
candidate for sending upstream as-is. If this works, I'll at least post
on update on the issue and let the community decide if this is worth
upstreaming.

Summary of the race condition: if the reschedule interval is shorter
than the duration of the scheduling loop, then a Sensor TI can
transition from RUNNING -> UP_FOR_RESCHEDULE -> QUEUED all before the
main scheduler loop can process the "task succeeded" event from the
executor.

If this happens, then the scheduler will see an event saying
"task succeeded" for a TI that's in state QUEUED, so it will mark the
task as failed and the user will see an error in the logs like:

    Executor reports task instance <TaskInstance: XXX [queued]> finished
    (success) although the task says its QUEUED. Was the task killed
    externally?

Instead, we should check and see if the TaskInstance is a sensor that
has been rescheduled. If it is, then this situation is expected and we
can just ignore it.

Here's the sequence of events in a little more detail:

1. [DagFileProcessorAgent] process_file() moves TI from NONE -> SCHEDULED
2. [SchedulerJob] _validate_and_run() -> _process_and_execute() -> _execute_task_instances(): moves TI from SCHEDULED -> QUEUED, sends to executor
3. [SchedulerJob] _validate_and_run() -> _process_executor_events() finishes without seeing any relevant events, scheduling loop restarts
4. [worker] Executor pokes, sets TI state to UP_FOR_RESCHEDULE and sends TI.event=SUCCESS
5. TI's reschedule interval expires, now eligble for rescheduling
6. [DagFileProcessorAgent] process_file() moves TI from UP_FOR_RESCHEDULE -> SCHEDULED
7. [SchedulerJob] SchedulerJob._validate_and_run() -> _process_and_execute() -> _execute_task_instances() -> moves TI from SCHEDULED -> QUEUED, sends to executor
8. [SchedulerJob] _validate_and_run() -> _process_executor_events() finally sees TI.event=SUCCESS from the previous attempt, but now TI.state == QUEUED. Oh no!